+++ R/hac_covariance.R
@@ -10,12 +10,11 @@
 #' @name HACcovariance
 #' @description This function estimates the long-run covariance matrix of a multivariate
 #' centred time series accounting for heteroskedasticity and autocorrelation
-#' using the Newey-West estimator.
-#' The number of lags is selected using the Newey-West (1994)
-#' <doi:10.2307/2297912> plug-in procedure, where
+#' using the Newey-West (1994)
+#' <doi:10.2307/2297912> estimator.
+#' The number is selected using the Newey-West plug-in procedure, where
 #' `n_lags = 4 * (n_observations/100)^(2/9)`.
-#' The function allows to internally prewhiten the series by fitting a VAR(1),
-#' i.e., a vector autoregressive model of order 1.
+#' The function allows to internally prewhiten the series by fitting a VAR(1).
 #' All the details can be found in Newey-West (1994)
 #' <doi:10.2307/2297912>.
 #'