+++ R/oracle_tfrp.R
@@ -7,7 +7,8 @@
 #' @title Oracle tradable factor risk premia.
 #'
 #' @name OracleTFRP
-#' @description Computes Oracle tradable factor risk premia from data on
+#' @description Computes Oracle tradable factor risk premia of
+#' Quaini-Trojani-Yuan (2023) <doi:10.2139/ssrn.4574683> from data on
 #' `K` factors `F = [F_1,...,F_K]'` and test asset excess returns `R`:
 #' `OTFRP = argmin_x ||TFRP - x||_2^2 + tau * sum_{k=1}^K w_k * |x_k|`,
 #' where `TFRP` is the tradable factor risk premia estimator, `tau > 0` is a