+++ src/RcppExports.cpp
@@ -65,6 +65,20 @@ BEGIN_RCPP
     return rcpp_result_gen;
 END_RCPP
 }
+// HJMisspecificationDistanceCpp
+Rcpp::List HJMisspecificationDistanceCpp(const arma::mat& returns, const arma::mat& factors, const double ci_coverage, const bool hac_prewhite);
+RcppExport SEXP _intrinsicFRP_HJMisspecificationDistanceCpp(SEXP returnsSEXP, SEXP factorsSEXP, SEXP ci_coverageSEXP, SEXP hac_prewhiteSEXP) {
+BEGIN_RCPP
+    Rcpp::RObject rcpp_result_gen;
+    Rcpp::RNGScope rcpp_rngScope_gen;
+    Rcpp::traits::input_parameter< const arma::mat& >::type returns(returnsSEXP);
+    Rcpp::traits::input_parameter< const arma::mat& >::type factors(factorsSEXP);
+    Rcpp::traits::input_parameter< const double >::type ci_coverage(ci_coverageSEXP);
+    Rcpp::traits::input_parameter< const bool >::type hac_prewhite(hac_prewhiteSEXP);
+    rcpp_result_gen = Rcpp::wrap(HJMisspecificationDistanceCpp(returns, factors, ci_coverage, hac_prewhite));
+    return rcpp_result_gen;
+END_RCPP
+}
 // ChenFang2019BetaRankTestCpp
 Rcpp::List ChenFang2019BetaRankTestCpp(const arma::mat& returns, const arma::mat& factors, const unsigned int n_bootstrap, const double target_level_kp2006_rank_test);
 RcppExport SEXP _intrinsicFRP_ChenFang2019BetaRankTestCpp(SEXP returnsSEXP, SEXP factorsSEXP, SEXP n_bootstrapSEXP, SEXP target_level_kp2006_rank_testSEXP) {
@@ -89,20 +103,6 @@ BEGIN_RCPP
     Rcpp::traits::input_parameter< const arma::mat& >::type factors(factorsSEXP);
     Rcpp::traits::input_parameter< const double >::type target_level(target_levelSEXP);
     rcpp_result_gen = Rcpp::wrap(IterativeKleibergenPaap2006BetaRankTestCpp(returns, factors, target_level));
-    return rcpp_result_gen;
-END_RCPP
-}
-// HJMisspecificationTestCpp
-Rcpp::List HJMisspecificationTestCpp(const arma::mat& returns, const arma::mat& factors, const double sqhj_distance_null_value, const bool hac_prewhite);
-RcppExport SEXP _intrinsicFRP_HJMisspecificationTestCpp(SEXP returnsSEXP, SEXP factorsSEXP, SEXP sqhj_distance_null_valueSEXP, SEXP hac_prewhiteSEXP) {
-BEGIN_RCPP
-    Rcpp::RObject rcpp_result_gen;
-    Rcpp::RNGScope rcpp_rngScope_gen;
-    Rcpp::traits::input_parameter< const arma::mat& >::type returns(returnsSEXP);
-    Rcpp::traits::input_parameter< const arma::mat& >::type factors(factorsSEXP);
-    Rcpp::traits::input_parameter< const double >::type sqhj_distance_null_value(sqhj_distance_null_valueSEXP);
-    Rcpp::traits::input_parameter< const bool >::type hac_prewhite(hac_prewhiteSEXP);
-    rcpp_result_gen = Rcpp::wrap(HJMisspecificationTestCpp(returns, factors, sqhj_distance_null_value, hac_prewhite));
     return rcpp_result_gen;
 END_RCPP
 }
@@ -196,9 +196,9 @@ static const R_CallMethodDef CallEntries[] = {
     {"_intrinsicFRP_GKRFactorScreeningCpp", (DL_FUNC) &_intrinsicFRP_GKRFactorScreeningCpp, 4},
     {"_intrinsicFRP_HACCovarianceMatrixCpp", (DL_FUNC) &_intrinsicFRP_HACCovarianceMatrixCpp, 2},
     {"_intrinsicFRP_HACVarianceCpp", (DL_FUNC) &_intrinsicFRP_HACVarianceCpp, 2},
+    {"_intrinsicFRP_HJMisspecificationDistanceCpp", (DL_FUNC) &_intrinsicFRP_HJMisspecificationDistanceCpp, 4},
     {"_intrinsicFRP_ChenFang2019BetaRankTestCpp", (DL_FUNC) &_intrinsicFRP_ChenFang2019BetaRankTestCpp, 4},
     {"_intrinsicFRP_IterativeKleibergenPaap2006BetaRankTestCpp", (DL_FUNC) &_intrinsicFRP_IterativeKleibergenPaap2006BetaRankTestCpp, 3},
-    {"_intrinsicFRP_HJMisspecificationTestCpp", (DL_FUNC) &_intrinsicFRP_HJMisspecificationTestCpp, 4},
     {"_intrinsicFRP_OracleTFRPGCVCpp", (DL_FUNC) &_intrinsicFRP_OracleTFRPGCVCpp, 14},
     {"_intrinsicFRP_OracleTFRPCVCpp", (DL_FUNC) &_intrinsicFRP_OracleTFRPCVCpp, 12},
     {"_intrinsicFRP_OracleTFRPRVCpp", (DL_FUNC) &_intrinsicFRP_OracleTFRPRVCpp, 14},