+++ src/hj_misspecification_distance.cpp
@@ -1,49 +1,44 @@
 // Author Alberto Quaini
 
-#include "misspecification_tests.h"
+#include "hj_misspecification_distance.h"
 #include "hac_covariance.h"
 
-////////////////////////////////////
-//// HJMisspecificationTestCpp ////
+///////////////////////////////////////
+//// HJMisspecificationDistanceCpp ////
 
-Rcpp::List HJMisspecificationTestCpp(
+Rcpp::List HJMisspecificationDistanceCpp(
   const arma::mat& returns,
   const arma::mat& factors,
-  const double sqhj_distance_null_value,
+  const double ci_coverage,
   const bool hac_prewhite
 ) {
 
   // Calculate the HJ misspecification statistic and p-value.
-  return HJMisspecificationStatisticAndPvalueCpp(
+  return HJMisspecificationDistanceCpp(
     returns,
     factors,
     arma::cov(returns),
     arma::mean(returns).t(),
-    sqhj_distance_null_value,
+    ci_coverage,
     hac_prewhite
   );
 
 }
 
-/////////////////////////////////////////////////
-//// HJMisspecificationStatisticAndPvalueCpp ////
+///////////////////////////////////////
+//// HJMisspecificationDistanceCpp ////
 
-Rcpp::List HJMisspecificationStatisticAndPvalueCpp(
+Rcpp::List HJMisspecificationDistanceCpp(
   const arma::mat& returns,
   const arma::mat& factors,
   const arma::mat& variance_returns,
   const arma::vec& mean_returns,
-  const double sqhj_distance_null_value,
+  const double ci_coverage,
   const bool hac_prewhite
 ) {
 
-  // Perform singular value decomposition on the variance of returns.
-  arma::vec eigval;
-  arma::mat eigvec;
-  arma::eig_sym(eigval, eigvec, variance_returns);
-
-  // Compute different powers of variance_returns.
-  const arma::mat inv_var_ret = eigvec * arma::diagmat(1. / eigval) * eigvec.t();
+  // Compute the inverse variance of returns.
+  const arma::mat inv_var_ret = arma::inv_sympd(variance_returns);
 
   // Calculate required vectors and matrices to compure the KRS SDF coefficients.
   const arma::vec var_ret_inv_mean_ret = arma::solve(
@@ -66,7 +61,7 @@ Rcpp::List HJMisspecificationStatisticAndPvalueCpp(
   ) * var_ret_inv_mean_ret;
 
   // Calculate the squared Hansen-Jagannathan (HJ) distance.
-  const double sqhj_distance =
+  const double squared_distance =
     arma::dot(mean_returns, var_ret_inv_mean_ret) - arma::dot(
       mean_returns.t() * var_ret_inv_cov_ret_fac,
       krs_sdf_coefficients
@@ -76,26 +71,29 @@ Rcpp::List HJMisspecificationStatisticAndPvalueCpp(
   const arma::mat returns_centred = returns.each_row() - mean_returns.t();
   const arma::mat factors_centred = factors.each_row() - arma::mean(factors);
 
-  // Follow computation in eq. (61) Kan-Robotti (2008)
-  // <10.1016/j.jempfin.2008.03.003>.
+  // Follow computation in eq. (61) Kan-Robotti (2008) <10.1016/j.jempfin.2008.03.003>.
   const arma::vec u = returns_centred * (
     var_ret_inv_mean_ret - var_ret_inv_cov_ret_fac * krs_sdf_coefficients
   );
   const arma::vec y =  1. - factors_centred * krs_sdf_coefficients;
-  arma::vec q = 2. * u % y - arma::square(u) + sqhj_distance;
+  arma::vec q = 2. * u % y - arma::square(u) + squared_distance;
 
-  const double variance_q = HACVarianceCpp(q, hac_prewhite);
+  // Compute the HAC standard errors of q.
+  const double standard_errors_q = std::sqrt(HACVarianceCpp(q, hac_prewhite));
 
-  // Compute the squared standardized HJ test statistic.
-  const double statistic = static_cast<double>(returns.n_rows) * std::pow(
-    sqhj_distance - sqhj_distance_null_value, 2
-  ) / variance_q;
+  // Set the right quantile of the normal based on the ci_coverage.
+  const double right_quantile =
+    R::qnorm5((1. - ci_coverage) / 2., 0., 1., false, false);
+
+  // Compute the shift around squared_distance determining the confidence interval.
+  const double shift = right_quantile * standard_errors_q /
+    std::sqrt(static_cast<double>(returns.n_rows));
 
   // Return the test statistic and the corresponding p-value.
   return Rcpp::List::create(
-    Rcpp::Named("sqhj_distance") = sqhj_distance,
-    Rcpp::Named("statistic") = statistic,
-    Rcpp::Named("p-value") = R::pchisq(statistic, 1, false, false)
+    Rcpp::Named("squared_distance") = squared_distance,
+    Rcpp::Named("lower_bound") = squared_distance - shift,
+    Rcpp::Named("upper_bound") = squared_distance + shift
   );
 
 }